+++ curia-front-end/app.json
@@ -5,7 +5,7 @@
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/images/icon.png",
-    "scheme": "myapp",
+    "scheme": "curia",
     "userInterfaceStyle": "automatic",
     "newArchEnabled": true,
     "ios": {
