+++ curia-front-end/src/app/_layout.tsx
@@ -1,7 +1,7 @@
 import { DevToolsBubble } from "react-native-react-query-devtools";
 import * as Clipboard from "expo-clipboard";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import App from "./app";
+import { Stack } from "expo-router";
 
 export default function RootLayout() {
   const queryClient = new QueryClient();
@@ -17,7 +17,9 @@ export default function RootLayout() {
 
   return (
     <QueryClientProvider client={queryClient}>
-      <App />
+      <Stack>
+        <Stack.Screen name="index" options={{ headerTitle: "Curia" }} />
+      </Stack>
       <DevToolsBubble onCopy={onCopy} />
     </QueryClientProvider>
   );
