+++ curia-front-end/src/app/index.tsx
@@ -1,4 +1,4 @@
-import SearchResults from "./components/SearchResults";
+import SearchResults from "../components/SearchResults";
 
 export default function Index() {
   return <SearchResults />;
