+++ curia-front-end/src/components/CollectionObjectListItem.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { Artefact } from "../apis/api.class";
 import { StyleSheet, View } from "react-native";
-import { Link } from "expo-router";
 
 interface Props {
   item: Artefact;
@@ -17,28 +16,24 @@ export default function CollectionObjectListItem({ item }: Props) {
   } = item;
 
   return (
-    <View role="listitem">
-      <Link href={`/artefact/${item.localId}`} asChild>
-        <div style={styles.container}>
-          <div style={styles.listItemLeft}>
-            <img src={primaryThumbnailUrl} alt="" style={styles.image} />
-          </div>
-          <div style={styles.listItemRight}>
-            <h2>{title}</h2>
-            <p>
-              {maker && maker + ", "}
-              {objectDate}
-            </p>
-            <p>Source: {apiSource}</p>
-          </div>
-        </div>
-      </Link>
+    <View role="listitem" style={styles.listItem}>
+      <div style={styles.listItemLeft}>
+        <img src={primaryThumbnailUrl} alt="" style={styles.image} />
+      </div>
+      <div style={styles.listItemRight}>
+        <h2>{title}</h2>
+        <p>
+          {maker && maker + ", "}
+          {objectDate}
+        </p>
+        <p>Source: {apiSource}</p>
+      </div>
     </View>
   );
 }
 
 const styles = StyleSheet.create({
-  container: {
+  listItem: {
     margin: 10,
     padding: 10,
     borderColor: "black",
@@ -46,7 +41,6 @@ const styles = StyleSheet.create({
     borderStyle: "solid",
     flexDirection: "row",
     display: "flex",
-    cursor: "pointer",
   },
   listItemLeft: {
     flex: 1,
